Changed lots of longs to int in the DPMI and LDT code, and a few for int15(extmem). Now DPMI LDT entry allocation works for 64bit.
git-svn-id: https://dosemu.svn.sourceforge.net/svnroot/dosemu/trunk@1611 7758b857-130e-0410-ab9c-d8735a528d18

+++ src/base/async/int.c
@@ -626,11 +626,11 @@ static int int15(void)
     break;
 
   case 0x87: {
-    unsigned long *lp;
+    unsigned int *lp;
     unsigned long src_addr, dst_addr;
     unsigned long src_limit, dst_limit;
     unsigned int length;
-    lp = SEG_ADR((long*), es, si);
+    lp = SEG_ADR((int*), es, si);
     lp += 4;
     src_addr = (*lp >> 16) & 0x0000FFFF;
     src_limit = *lp & 0x0000FFFF;
